Address review pt.3
