+++ src/pipeline.ts
@@ -2,11 +2,7 @@ import chalk from "chalk";
 import fs from "fs";
 import { zip } from "lodash-es";
 import ms from "ms";
-import {
-	linkAllFilesInMetafile,
-	performAction,
-	performActions,
-} from "./action.js";
+import { performAction, performActions } from "./action.js";
 import {
 	ActionResult,
 	Decision,
@@ -14,18 +10,13 @@ import {
 	InjectionResult,
 	isAnyMatchedDecision,
 	SaveResult,
-	UNKNOWN_TRACKER,
 } from "./constants.js";
 import {
 	findPotentialNestedRoots,
 	findSearcheesFromAllDataDirs,
 } from "./dataFiles.js";
 import { db } from "./db.js";
-import {
-	assessCandidate,
-	assessCandidateCaching,
-	ResultAssessment,
-} from "./decide.js";
+import { assessCandidateCaching, ResultAssessment } from "./decide.js";
 import {
 	IndexerStatus,
 	updateIndexerStatus,
@@ -49,14 +40,11 @@ import {
 	SearcheeWithLabel,
 } from "./searchee.js";
 import {
-	findAllTorrentFilesInDir,
 	getInfoHashesToExclude,
 	getTorrentByCriteria,
 	getSimilarTorrentsByName,
 	indexNewTorrents,
 	loadTorrentDirLight,
-	parseInfoFromSavedTorrent,
-	parseTorrentFromFilename,
 	TorrentLocator,
 } from "./torrent.js";
 import {
@@ -65,17 +53,7 @@ import {
 	queryRssFeeds,
 	searchTorznab,
 } from "./torznab.js";
-import {
-	areMediaTitlesSimilar,
-	formatAsList,
-	getLogString,
-	isTruthy,
-	sanitizeInfoHash,
-	wait,
-} from "./utils.js";
-import { Metafile } from "./parseTorrent.js";
-import { getClient } from "./clients/TorrentClient.js";
-import { dirname } from "path";
+import { getLogString, isTruthy, wait } from "./utils.js";
 
 export interface Candidate {
 	guid: string;
@@ -96,27 +74,6 @@ interface FoundOnOtherSites {
 	searchedIndexers: number;
 	matches: number;
 }
-
-type FullMatches = [
-	SearcheeWithLabel,
-	Decision.MATCH | Decision.MATCH_SIZE_ONLY,
-][];
-type PartialMatches = [SearcheeWithLabel, Decision.MATCH_PARTIAL][];
-
-type InjectSummary = {
-	TOTAL: number;
-	INJECTED: number;
-	FULL_MATCHES: number;
-	PARTIAL_MATCHES: number;
-	BLOCKED: number;
-	ALREADY_EXISTS: number;
-	INCOMPLETE_CANDIDATES: number;
-	INCOMPLETE_SEARCHEES: number;
-	FAILED: number;
-	UNMATCHED: number;
-	FOUND_BAD_FORMAT: boolean;
-	FLAT_LINKING: boolean;
-};
 
 async function assessCandidates(
 	candidates: Candidate[],
@@ -403,7 +360,7 @@ export async function checkNewCandidateMatch(
 	return { decision, actionResult };
 }
 
-async function findAllSearchees(
+export async function findAllSearchees(
 	searcheeLabel: SearcheeLabel,
 ): Promise<SearcheeWithLabel[]> {
 	const { torrents, dataDirs, torrentDir } = getRuntimeConfig();
@@ -560,474 +517,4 @@ export async function scanRssFeeds() {
 		}
 		logger.info({ label: Label.RSS, message: "Scan complete" });
 	}
-}
-
-async function injectDecideStage(
-	meta: Metafile,
-	searchees: SearcheeWithLabel[],
-): Promise<{
-	fullMatches: FullMatches;
-	partialMatches: PartialMatches;
-	foundBlocked: boolean;
-}> {
-	const fullMatches: [
-		SearcheeWithLabel,
-		Decision.MATCH | Decision.MATCH_SIZE_ONLY,
-	][] = [];
-	const partialMatches: [SearcheeWithLabel, Decision.MATCH_PARTIAL][] = [];
-	let foundBlocked = false;
-	for (const searchee of searchees) {
-		const { decision } = await assessCandidate(meta, searchee, []);
-		if (!isAnyMatchedDecision(decision)) {
-			if (decision === Decision.BLOCKED_RELEASE) {
-				foundBlocked = true;
-			}
-			continue;
-		}
-
-		// If name or file names are not similar consider it a false positive
-		if (
-			!areMediaTitlesSimilar(searchee.title, meta.title) &&
-			!areMediaTitlesSimilar(searchee.title, meta.name) &&
-			!areMediaTitlesSimilar(searchee.name, meta.name) &&
-			!areMediaTitlesSimilar(searchee.name, meta.title) &&
-			!meta.files.some((metaFile) =>
-				searchee.files.some((searcheeFile) =>
-					areMediaTitlesSimilar(searcheeFile.name, metaFile.name),
-				),
-			)
-		) {
-			logger.warn({
-				label: Label.INJECT,
-				message: `Skipping potential false positive for ${getLogString(meta, chalk.bold.white)} from ${getLogString(searchee, chalk.bold.white)}`,
-			});
-			continue;
-		}
-
-		if (decision === Decision.MATCH) {
-			fullMatches.unshift([searchee, decision]);
-		} else if (decision === Decision.MATCH_SIZE_ONLY) {
-			fullMatches.push([searchee, decision]);
-		} else {
-			partialMatches.push([searchee, decision]);
-		}
-	}
-	fullMatches.sort(([searcheeA], [searcheeB]) => {
-		// Prefer torrent over data/virtual, only torrent knows if it's complete
-		if (searcheeA.infoHash && !searcheeB.infoHash) {
-			return -1;
-		}
-		if (!searcheeA.infoHash && searcheeB.infoHash) {
-			return 1;
-		}
-		return 0; // Should keep MATCH first within a searchee type
-	});
-	partialMatches.sort(([searcheeA], [searcheeB]) => {
-		// Prefer torrent/data over virtual, partials are always rechecked
-		if (
-			(searcheeA.infoHash || searcheeA.path) &&
-			!(searcheeB.infoHash || searcheeB.path)
-		) {
-			return -1;
-		}
-		if (
-			!(searcheeA.infoHash || searcheeA.path) &&
-			(searcheeB.infoHash || searcheeB.path)
-		) {
-			return 1;
-		}
-		return searcheeB.files.length - searcheeA.files.length; // Prefer more files
-	});
-	return { fullMatches, partialMatches, foundBlocked };
-}
-
-async function injectInitialAction(
-	meta: Metafile,
-	fullMatches: FullMatches,
-	partialMatches: PartialMatches,
-	tracker: string,
-): Promise<{
-	injectionResult: InjectionResult;
-	matchedSearchee?: SearcheeWithLabel;
-	matchedDecision?: DecisionAnyMatch;
-	linkedNewFiles: boolean;
-}> {
-	let injectionResult = InjectionResult.FAILURE;
-	let matchedSearchee: SearcheeWithLabel | undefined;
-	let matchedDecision: DecisionAnyMatch | undefined;
-	let linkedNewFiles = false;
-	for (const [searchee, decision] of [...fullMatches, ...partialMatches]) {
-		if (
-			injectionResult === InjectionResult.TORRENT_NOT_COMPLETE &&
-			!searchee.infoHash
-		) {
-			continue; // Data/virtual searchee doesn't know if torrent is complete
-		}
-		const res = await performAction(meta, decision, searchee, tracker);
-		const result = res.actionResult;
-		if (res.linkedNewFiles) {
-			linkedNewFiles = true;
-		}
-		if (
-			injectionResult === InjectionResult.SUCCESS ||
-			result === InjectionResult.FAILURE ||
-			result === SaveResult.SAVED
-		) {
-			continue;
-		}
-		if (result === InjectionResult.ALREADY_EXISTS) {
-			injectionResult = result;
-			continue;
-		}
-		if (result === InjectionResult.TORRENT_NOT_COMPLETE) {
-			if (injectionResult !== InjectionResult.ALREADY_EXISTS) {
-				injectionResult = result;
-				matchedSearchee = searchee;
-				matchedDecision = decision;
-			}
-			continue;
-		}
-		injectionResult = InjectionResult.SUCCESS;
-		matchedSearchee = searchee;
-		matchedDecision = decision;
-	}
-	return {
-		injectionResult,
-		matchedSearchee,
-		matchedDecision,
-		linkedNewFiles,
-	};
-}
-
-async function injectFromStalledTorrent(
-	meta: Metafile,
-	fullMatches: FullMatches,
-	partialMatches: PartialMatches,
-	tracker: string,
-	injectionResult: InjectionResult,
-	progress: string,
-	filePathLog: string,
-): Promise<boolean> {
-	let linkedNewFiles = false;
-	let inClient = (await getClient().isTorrentComplete(meta.infoHash)).isOk();
-	let injected = false;
-	for (const [searchee, decision] of [...fullMatches, ...partialMatches]) {
-		const linkedFilesRootResult = await linkAllFilesInMetafile(
-			searchee,
-			meta,
-			tracker,
-			decision,
-			{ onlyCompleted: false },
-		);
-		const linkResult = linkedFilesRootResult.isOk()
-			? linkedFilesRootResult.unwrap()
-			: null;
-		if (linkResult && linkResult.linkedNewFiles) {
-			linkedNewFiles = true;
-		}
-		if (!inClient) {
-			if (linkedFilesRootResult.isOk()) {
-				const destinationDir = dirname(linkResult!.contentPath);
-				const result = await getClient().inject(
-					meta,
-					searchee,
-					Decision.MATCH_PARTIAL, // Should always be considered partial
-					destinationDir,
-				);
-				// result is only SUCCESS or FAILURE here but still log original injectionResult
-				if (result === InjectionResult.SUCCESS) {
-					logger.info({
-						label: Label.INJECT,
-						message: `${progress} Injected ${filePathLog} using stalled source, you will need to resume or remove from client - ${chalk.green(injectionResult)}`,
-					});
-					inClient = true;
-					injected = true;
-				} else {
-					logger.warn({
-						label: Label.INJECT,
-						message: `${progress} Failed to inject ${filePathLog} using stalled source - ${chalk.yellow(injectionResult)}`,
-					});
-				}
-			} else {
-				logger.warn({
-					label: Label.INJECT,
-					message: `${progress} Failed to link files for ${filePathLog}, ${linkedFilesRootResult.unwrapErr()} - ${chalk.yellow(injectionResult)}`,
-				});
-			}
-		}
-	}
-	if (inClient && !injected) {
-		if (linkedNewFiles) {
-			logger.info({
-				label: Label.INJECT,
-				message: `${progress} Rechecking ${filePathLog} as new files were linked - ${chalk.green(injectionResult)}`,
-			});
-			await getClient().recheckTorrent(meta.infoHash);
-		} else {
-			logger.warn({
-				label: Label.INJECT,
-				message: `${progress} No new files linked for ${filePathLog}, resume or remove from client - ${chalk.yellow(injectionResult)}`,
-			});
-		}
-	}
-	return injected;
-}
-
-function injectSummary(summary: InjectSummary): void {
-	const incompleteMsg = `${chalk.bold.yellow(summary.ALREADY_EXISTS)} existed in client${
-		summary.INCOMPLETE_CANDIDATES
-			? chalk.dim(` (${summary.INCOMPLETE_CANDIDATES} were incomplete)`)
-			: ""
-	}`;
-	const resultMsg = formatAsList(
-		[
-			`Injected ${chalk.bold.green(summary.INJECTED)}/${chalk.bold.white(summary.TOTAL)} torrents`,
-			summary.FULL_MATCHES &&
-				`${chalk.bold.green(summary.FULL_MATCHES)} were full matches`,
-			summary.PARTIAL_MATCHES &&
-				`${chalk.bold.yellow(summary.PARTIAL_MATCHES)} were partial matches`,
-			summary.INCOMPLETE_SEARCHEES &&
-				`${chalk.bold.yellow(summary.INCOMPLETE_SEARCHEES)} had incomplete sources`,
-			summary.ALREADY_EXISTS && incompleteMsg,
-			summary.BLOCKED &&
-				`${chalk.bold.yellow(summary.BLOCKED)} were possibly blocklisted`,
-			summary.FAILED &&
-				`${chalk.bold.red(summary.FAILED)} failed to inject`,
-			summary.UNMATCHED &&
-				`${chalk.bold.red(summary.UNMATCHED)} had no matches`,
-		].filter(isTruthy),
-		{ sort: false, type: "unit" },
-	);
-	logger.info({ label: Label.INJECT, message: chalk.cyan(resultMsg) });
-
-	if (summary.UNMATCHED > 0) {
-		logger.info({
-			label: Label.INJECT,
-			message: `Use "${chalk.bold.white("cross-seed diff")}" to get the reasons two torrents are not considered matches`,
-		});
-	}
-
-	if (summary.FOUND_BAD_FORMAT && !summary.FLAT_LINKING) {
-		logger.warn({
-			label: Label.INJECT,
-			message: `Some torrents could be linked to linkDir/${UNKNOWN_TRACKER} - follow .torrent naming format in the docs to avoid this`,
-		});
-	}
-}
-
-export async function injectSavedTorrents() {
-	const { flatLinking, injectDir, linkDir, outputDir } = getRuntimeConfig();
-	const targetDir = injectDir ?? outputDir;
-	const targetDirLog = chalk.bold.magenta(targetDir);
-
-	const dirContents = await findAllTorrentFilesInDir(targetDir);
-	if (dirContents.length === 0) {
-		logger.info({
-			label: Label.INJECT,
-			message: `No torrent files found to inject in ${targetDirLog}`,
-		});
-		return;
-	}
-	logger.info({
-		label: Label.INJECT,
-		message: `Found ${chalk.bold.white(dirContents.length)} torrent file(s) to inject in ${targetDirLog}`,
-	});
-
-	const searchees = await findAllSearchees(Label.INJECT);
-
-	const toDelete = new Set<string>();
-
-	// Usually source got deleted or partial injection never completes
-	function shouldCleanUpTorrent(torrentFilePath: string) {
-		return fs.statSync(torrentFilePath).mtimeMs < Date.now() - ms("1 week");
-	}
-	function getTorrentFilePathLog(torrentFilePath: string) {
-		return chalk.bold.magenta(
-			torrentFilePath.replace(
-				/\[([a-z0-9]{40})].torrent$/i,
-				(match, hash) => match.replace(hash, sanitizeInfoHash(hash)),
-			),
-		);
-	}
-
-	const summary: InjectSummary = {
-		TOTAL: dirContents.length,
-		INJECTED: 0,
-		FULL_MATCHES: 0,
-		PARTIAL_MATCHES: 0,
-		BLOCKED: 0,
-		ALREADY_EXISTS: 0,
-		INCOMPLETE_CANDIDATES: 0,
-		INCOMPLETE_SEARCHEES: 0,
-		FAILED: 0,
-		UNMATCHED: 0,
-		FOUND_BAD_FORMAT: false,
-		FLAT_LINKING: flatLinking,
-	};
-	let count = 0;
-	for (const torrentFilePath of dirContents) {
-		const progress = chalk.blue(`(${++count}/${dirContents.length})`);
-		const filePathLog = getTorrentFilePathLog(torrentFilePath);
-		if (shouldCleanUpTorrent(torrentFilePath)) {
-			toDelete.add(torrentFilePath);
-		}
-		let meta: Metafile;
-		try {
-			meta = await parseTorrentFromFilename(torrentFilePath);
-		} catch (e) {
-			logger.error({
-				label: Label.INJECT,
-				message: `${progress} Failed to parse ${filePathLog}`,
-			});
-			logger.debug(e);
-			continue;
-		}
-		const metaLog = getLogString(meta, chalk.bold.white);
-
-		const torrentNameInfo =
-			await parseInfoFromSavedTorrent(torrentFilePath);
-		const tracker = torrentNameInfo?.tracker ?? UNKNOWN_TRACKER;
-		if (tracker === UNKNOWN_TRACKER) {
-			summary.FOUND_BAD_FORMAT = true;
-		}
-
-		const { fullMatches, partialMatches, foundBlocked } =
-			await injectDecideStage(meta, searchees);
-		if (fullMatches.length === 0 && partialMatches.length === 0) {
-			if (foundBlocked) {
-				logger.info({
-					label: Label.INJECT,
-					message: `${progress} ${metaLog} ${chalk.yellow("possibly blocklisted")}: ${filePathLog}`,
-				});
-				summary.BLOCKED++;
-			} else {
-				logger.info({
-					label: Label.INJECT,
-					message: `${progress} ${metaLog} ${chalk.red("has no matches")}: ${filePathLog}`,
-				});
-				summary.UNMATCHED++;
-			}
-			continue;
-		}
-
-		const {
-			injectionResult,
-			matchedSearchee,
-			matchedDecision,
-			linkedNewFiles,
-		} = await injectInitialAction(
-			meta,
-			fullMatches,
-			partialMatches,
-			tracker,
-		);
-		if (injectionResult === InjectionResult.FAILURE) {
-			logger.error({
-				label: Label.INJECT,
-				message: `${progress} Failed to inject ${filePathLog} - ${chalk.red(injectionResult)}`,
-			});
-			summary.FAILED++;
-			toDelete.delete(torrentFilePath);
-			continue;
-		}
-		if (injectionResult === InjectionResult.TORRENT_NOT_COMPLETE) {
-			if (
-				!linkDir ||
-				fs.statSync(torrentFilePath).mtimeMs >= Date.now() - ms("1 day")
-			) {
-				// Normal case where source is likely still downloading
-				logger.warn({
-					label: Label.INJECT,
-					message: `${progress} Unable to inject ${filePathLog} - ${chalk.yellow(injectionResult)}`,
-				});
-			} else {
-				// Since source is stalled, add to client paused so user can resume later if desired
-				// Try linking all possible matches as they may have different files
-				await injectFromStalledTorrent(
-					meta,
-					fullMatches,
-					partialMatches,
-					tracker,
-					injectionResult,
-					progress,
-					filePathLog,
-				);
-			}
-			summary.INCOMPLETE_SEARCHEES++;
-			continue;
-		}
-		const result = await getClient().isTorrentComplete(meta.infoHash);
-		let isComplete = result.isOk() ? result.unwrap() : false;
-		if (isComplete) {
-			toDelete.add(torrentFilePath);
-		}
-		if (injectionResult === InjectionResult.ALREADY_EXISTS) {
-			if (linkedNewFiles) {
-				logger.info({
-					label: Label.INJECT,
-					message: `${progress} Rechecking ${filePathLog} as new files were linked - ${chalk.green(injectionResult)}`,
-				});
-				await getClient().recheckTorrent(meta.infoHash);
-			} else if (fullMatches.length && !isComplete) {
-				logger.info({
-					label: Label.INJECT,
-					message: `${progress} Rechecking ${filePathLog} as it's not complete but has all files - ${chalk.green(injectionResult)}`,
-				});
-				isComplete = true;
-				await getClient().recheckTorrent(meta.infoHash);
-				toDelete.add(torrentFilePath); // Prevent infinite recheck in rare case of corrupted cross seed
-			} else {
-				logger.warn({
-					label: Label.INJECT,
-					message: `${progress} Unable to inject ${filePathLog} - ${chalk.yellow(injectionResult)}${isComplete ? "" : " (incomplete)"}`,
-				});
-			}
-			summary.ALREADY_EXISTS++;
-			summary.INCOMPLETE_CANDIDATES += isComplete ? 0 : 1;
-			continue;
-		}
-		logger.info({
-			label: Label.INJECT,
-			message: `${progress} Injected ${filePathLog} - ${chalk.green(injectionResult)}`,
-		});
-		sendResultsNotification(matchedSearchee!, [
-			[
-				{ decision: matchedDecision!, metafile: meta },
-				tracker,
-				injectionResult,
-			],
-		]);
-		summary.INJECTED++;
-		if (matchedDecision! === Decision.MATCH_PARTIAL) {
-			summary.PARTIAL_MATCHES++;
-		} else {
-			summary.FULL_MATCHES++;
-		}
-	}
-
-	for (const torrentFilePath of toDelete) {
-		const filePathLog = getTorrentFilePathLog(torrentFilePath);
-		try {
-			if (shouldCleanUpTorrent(torrentFilePath)) {
-				logger.warn({
-					label: Label.INJECT,
-					message: `Deleting ${filePathLog} as it has failed to inject for too long`,
-				});
-			} else {
-				logger.info({
-					label: Label.INJECT,
-					message: `Deleting ${filePathLog} as it's in client and complete`,
-				});
-			}
-			fs.unlinkSync(torrentFilePath);
-		} catch (e) {
-			logger.error({
-				label: Label.INJECT,
-				message: `Failed to delete ${filePathLog}`,
-			});
-			logger.debug(e);
-		}
-	}
-
-	injectSummary(summary);
 }
